relatorios/CPqD_Antifraude_Relatorios_Usuario/requirements.txt
oracledb==2.4.1
python-dotenv==1.0.1
